Moved Button class to separate file
